Mise en place RouteActionProcess pour recherche artiste par nom
