+++ yake_back/Composants/Artiste/Artiste.process.js
@@ -10,9 +10,24 @@ app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({extended : true}));
 app.use(session({secret: 'todotopsecret'}))
 
-const Routes = require('./Artiste.route');
 const Actions = require('./Artiste.action');
 const Model = require('./Artiste.model');
+
+exports.findArtisteByName = function(genres,name){
+	return new Promise(function(resolve,reject){
+		if(name===undefined ||name.length===0){
+			reject(404);
+		}
+		else{
+		Actions.getArtisteByName(genres,name)
+		.then(artistes=>{
+				artistes = Actions.formatageBlocsArtistepourEnvoi(artistes);
+				resolve(artistes);
+			})
+		.catch((err)=>{reject(404);});
+		}
+	});
+}
 
 // app.get(, function(req, res) {
 // })
